+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # CHANGELOG for @dictadata/storage-node
 
 - version 0.9.69
-  - bug fix: coerce smt to strings in tracts entries
+  - bug fix: coerce smt to strings in cortex entries
   - updated: @dictadata/storage-junctions@0.9.94
 - version 0.9.68
-  - feature: Tracts storage
+  - feature: Cortex storage
   - updated: @dictadata/storage-junctions@0.9.93
 - version 0.9.67
   - bug fix: have process.env.HOMEPATH take precedence over os.homedir()